pyproject.toml
[tool.poetry]
name = "anaya"
version = "0.1.0"
description = "Anaya is a Content Engine that specializes in analyzing and comparing multiple PDF documents. It uses Retrieval Augmented Generation (RAG) techniques to effectively retrieve, assess, and generate insights from the documents."
authors = ["Shreyas Waghmode"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.36.0"
llama-index = "^0.10.51"
rapidfuzz = "^3.9.4"
langchain-community = "^0.2.6"
langchain = "^0.2.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
